# challenges/challenge-1-web-api/python-fastapi/openapi.yaml
openapi: "3.0.3"
info:
  title: EDP Field Ops — Task Manager API
  version: "1.0.0"
  description: |
    REST API for managing field operations work orders at EDP.
    Track outages, grid maintenance, solar installations, EV charging deployments and safety inspections.
servers:
  - url: http://127.0.0.1:8000
    description: Local development server

# ── Security ───────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable schemas ────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [viewer, user, admin]
    TaskStatus:
      type: string
      enum: [reported, assigned, in_progress, resolved, closed]
    TaskPriority:
      type: string
      enum: [low, medium, high, critical]
    TaskCategory:
      type: string
      enum:
        - grid_maintenance
        - outage_response
        - meter_inspection
        - solar_installation
        - ev_charging
        - customer_request
        - safety_inspection

    # Auth
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    TaskCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
          default: reported
        priority:
          $ref: "#/components/schemas/TaskPriority"
          default: medium
        category:
          $ref: "#/components/schemas/TaskCategory"
          default: grid_maintenance
        location:
          type: string
          maxLength: 200
          nullable: true
        asset_id:
          type: string
          maxLength: 50
          nullable: true
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        category:
          $ref: "#/components/schemas/TaskCategory"
        location:
          type: string
          maxLength: 200
          nullable: true
        asset_id:
          type: string
          maxLength: 50
          nullable: true
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        category:
          $ref: "#/components/schemas/TaskCategory"
          nullable: true
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryResponse"
        location:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # Categories
...
    AuditLogResponse:
      type: object
      properties:
        id:
          type: integer
        task_id:
          type: integer
        action:
          type: string
          enum: [create, update, delete]
        changed_by:
          type: string
        field:
          type: string
          nullable: true
          description: The task field that changed (null for create/delete entries)
        old_value:
          type: string
          nullable: true
        new_value:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # Errors
    HTTPError:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

  # ── Reusable responses ──────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Not authenticated"
    Forbidden:
      description: Authenticated but insufficient role or not the owner
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "You do not have permission to modify this task"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Task not found"
    UnprocessableEntity:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
...
security:
  - BearerAuth: []

# ============================================================================
# Paths
# ============================================================================
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: List all tasks
      description: All authenticated roles see all tasks. Supports filtering and pagination.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - name: category
          in: query
          description: Filter by legacy single category
          schema:
            $ref: "#/components/schemas/TaskCategory"
        - name: location
          in: query
          schema:
            type: string
        - name: q
          in: query
          description: Full-text search in title and description
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [tasks]
      summary: Create a new task
      description: Requires `user` or `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
            examples:
              outage:
                summary: Outage response
                value:
                  title: "Power outage — Zona Industrial Setúbal"
                  description: "Full blackout affecting 3 industrial units."
                  status: reported
                  priority: critical
                  category: outage_response
                  location: "Zona Industrial Setúbal, Lote 14"
                  asset_id: TR-9021-STB
              solar:
                summary: Solar installation
                value:
                  title: "Residential solar install — Lisboa Norte"
                  status: assigned
                  priority: medium
                  category: solar_installation
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [tasks]
      summary: Get a task by ID
      description: All authenticated roles can view any task.
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [tasks]
      summary: Replace all task fields
      description: Full replacement. Requires `user` or `admin` role. Owner or admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    patch:
      tags: [tasks]
      summary: Partially update a task
      description: Only supplied fields are updated. Requires `user` or `admin` role. Owner or admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
            examples:
              escalate:
                summary: Escalate priority
                value:
                  status: assigned
                  priority: critical
              resolve:
                summary: Mark in progress
                value:
                  status: in_progress
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    delete:
      tags: [tasks]
      summary: Delete a task
      description: Requires `user` or `admin` role. Owner or admin only.
      responses:
        "204":
          description: Task deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tasks/{task_id}/audit:
    get:
      tags: [audit]
      summary: Get the audit trail for a task
      description: |
        Returns every recorded mutation (create, update, delete) for a task, newest first.
        Update entries have one row per changed field, each with `old_value` / `new_value`.
        Delete entries survive task deletion (no CASCADE on `task_id`).
        All authenticated roles can read the audit log.
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogResponse"
              example:
                - id: 3
                  task_id: 2
                  action: update
                  changed_by: joana@edp.pt
                  field: asset_id
                  old_value: "123"
                  new_value: TR-9021-STB
                  timestamp: "2026-04-16T13:23:35Z"
                - id: 1
                  task_id: 2
                  action: create
                  changed_by: joana@edp.pt
                  field: null
                  old_value: null
                  new_value: null
                  timestamp: "2026-04-16T10:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Categories ────────────────────────────────────────────────────────────
...
  /admin/tasks:
    get:
      tags: [admin]
      summary: List all tasks (admin view)
      description: Like `GET /tasks` but scoped across all users. Can filter by `user_id`. Requires `admin` role.
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          schema:
            $ref: "#/components/schemas/TaskPriority"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/TaskCategory"
        - name: user_id
          in: query
          description: Filter by owner user ID
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Paginated task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      description: Requires `admin` role.
      responses:
        "200":
          description: User list ordered by ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{user_id}/role:
    patch:
      tags: [admin]
      summary: Change a user's role
      description: |
        Requires `admin` role.
        Admins cannot change their own role (`400`).
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
            example:
              role: user
      responses:
        "200":
          description: User role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Admin cannot change their own role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: You cannot change your own role
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

# ── Tags ──────────────────────────────────────────────────────────────────
...